feat: add registerIcon
