+++ apps/cli/src/template/vtable/nextjs-demo/src/lib/registerIcon.ts
@@ -1,0 +1,10 @@
+import icons from "@/config/icons";
+import * as VTable from "@visactor/vtable";
+
+export function registerIcon() {
+  icons.forEach((icon) => {
+    VTable.register.icon(icon.name, icon.spec);
+  });
+}
+
+registerIcon();
